created action file for application modal and added open notification for now
