+++ src/state/reducers/applicationsReducer.js
@@ -1,0 +1,9 @@
+// import { handleApplication } from "../actions/applications/handleApplications";
+
+// const initialState = {
+//     currentApplication: {
+//         role: '',
+//         status: 'pending'
+//     },
+//     profileId: ''
+// };
